requirements.txt
yfinance>=0.2.36
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
python-telegram-bot>=20.7
pytz>=2024.1
psutil>=5.9.8

